pyproject.toml
[project]
name = "aboutdogs"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrew Skow",email = "anjeie24@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djhtml (>=3.0.10,<4.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ipython (>=9.7.0,<10.0.0)",
    "django-phonenumber-field[phonenumbers] (>=8.4.0,<9.0.0)",
    "django-crispy-forms (>=2.5,<3.0)",
    "crispy-bootstrap5 (>=2025.6,<2026.0)",
    "certifi (>=2025.11.12,<2026.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
